Update common.py
Fixing errmeddatacol assigned but never used in get_bins(). This whole section for estimating uncertainties of pixels with replaced values was a bit of a mess. The uncertainties are now estimated very conservatively, and may be overestiamated. Given that these pixels formerly had cosmic rays in them, it shouldn't be too much trouble. Cleanup of this section and removal of commented lines will happen later after testing.