motes.py / save_fits(): removing "INTERP_KIND" keyword from motesparams
"INTERP_KIND" used to refer to the type of interpolation used for defining the limits of the extraction aperture. This has been hard coded to linear interpolation for a while, and "INTERP_KIND" is now a useless parameter in motesparams.txt. The piece of motes.py being removed is the only place it is referenced in all of the codebase, so removing it does no harm.